requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pymupdf
sentence-transformers
python-dotenv